Decompose engine.ts into focused modules
Extract pure functions from the 813-line monolith:
- engine/returns.ts — return schema validation, type checking
- engine/wait.ts — duration parsing, wait conditions, timeout logic
- engine/helpers.ts — toNodeInfo, cloneContext

Core class remains at 662 lines — subgraph push/pop stays coupled
to class state. All 279 tests pass, clean compile.
