+++ src/engine/index.ts
@@ -1,0 +1,1 @@
+export { GraphEngine } from "./engine.js";
